Update javascript.js
Okay, this is the quotation form handler. It includes features such as, like it gathers form input values. And also, it constructs a report based on the input that you have put in. Also, it has logic to display, like confirmation message. And it also resets, right? It is a quote form reset. It uses  the getElementById to do so as well, to work everything out, basically. Yeah, so that's the quotation form oh and it calculates the estimated price based off the time of goods and distance