Making pages more dynamic
Okay so some major changes occurred. I implemented a lot of hover movements as well as fades in order to make all the pages more dynamic so I had to make changes on my JavaScript as well as my CSS and also implement the different classes in all the pages in order for the fades and models to work properly. Now every single page is way more dynamic. Yes that is about it.